requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain-google-vertexai ##embedding
langchain
streamlit
langchain_community
python-dotenv
pypdf